refactored viewDidLoad by abstracting away previous navigation bar/item configuration and table view delegate/data source into their own separate functions and calling these functions within viewDidLoad
adjusted the weight of set variable to be a double instead of an Int

adjusted the addSetToExercise objc function to remove trailing/leading whitespace

+++ WorkoutPlannerApp/Model/Exercise/ExerciseSetModel.swift
@@ -10,7 +10,7 @@ import Foundation
 
 struct ExerciseSetModel: Codable {
 
-    var weightOfSet: Int = 0
+    var weightOfSet: Double = 0
     var numberOfReps: Int = 0
     
 }
